requirements.txt
beautifulsoup4==4.12.2
chromadb==0.3.23
numpy==1.22.3
openai==0.27.2
pandas==2.0.1
pandas_stubs==1.2.0.53
python-dotenv==1.0.0
requests==2.28.2
rich==13.3.5
streamlit==1.21.0
tiktoken==0.3.3
